Add AudioManager & beberapa music + SFX
