+++ Till and Thrive/Assets/Music/C418 - Haggstrom - Minecraft Volume Alpha.mp3.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 72b31abdf4359464c99f711acbb4942d
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 7
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
