+++ Till and Thrive/Assets/Music/[Video Soundtrack] Farm Boy [FINAL FANTASY VII].mp3.meta
@@ -1,0 +1,23 @@
+fileFormatVersion: 2
+guid: 2aeaa006a5c64dc4e985fb8072f645e8
+AudioImporter:
+  externalObjects: {}
+  serializedVersion: 7
+  defaultSettings:
+    serializedVersion: 2
+    loadType: 0
+    sampleRateSetting: 0
+    sampleRateOverride: 44100
+    compressionFormat: 1
+    quality: 1
+    conversionMode: 0
+    preloadAudioData: 0
+  platformSettingOverrides: {}
+  forceToMono: 0
+  normalize: 1
+  loadInBackground: 0
+  ambisonic: 0
+  3D: 1
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
