+++ Till and Thrive/Assets/Scripts/Main Menu/MainMenuScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c9207c46df2d37b48ae218f55b8143f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
