+++ Till and Thrive/Assets/Scripts/MainGame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c00cc845be402a34d9a9cec81a90c567
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
