+++ Till and Thrive/Assets/Scripts/MainGame/BackgroundMusic/MusicLoop.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ca763e9230e17684ea5e825eee830db1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
