+++ Till and Thrive/Assets/Scripts/MainGame/MapBorders.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 724c79e3ef7d3fd47975a7c8e9f00c8e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
