+++ Till and Thrive/Assets/Scripts/MainGame/Player/PlayerMoney.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb83a783330cd4445ab92ffb1daa48ec
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
