--- Till and Thrive/Assets/Scripts/MapBorders/Border.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3cc90d3dfcb361944a5511504e1f376d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
