--- Till and Thrive/Assets/Scripts/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1da1f860f6fd226459a4476535847202
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
