--- Till and Thrive/Assets/Scripts/Player/CameraMovement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fd11799fdee10a04290873e4b970a203
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
