+++ Till and Thrive/Assets/Sound Effect.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c5d960c6601bc6f44a56562161ce3fc2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
